+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.10.2 - ?
 ===================
 - Add line numbers support (absolute and relative)
+- Add Delete key support in insert mode (DeleteCharForward)
 
 Released
 --------